Commit 12
Todo completado en su mayoria, sinceramente faltan algunas cosas, puede que se agreguen luego pero no se si serán validas, de todos modos la base de todo en su mayoría se encuentra bien y está listo

+++ src/main/java/Tienda/Tienda/db/IUserRepository.java
@@ -4,7 +4,7 @@ import Tienda.Tienda.entities.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IUserRepository extends JpaRepository<Usuario, Long> {
+public interface IUserRepository extends JpaRepository<Usuario, Integer> {
     
     Usuario findByUsername(String Username);
